finterface: Get the return type right.

+++ src/lokios.1/net/mock/finterface.h
@@ -19,7 +19,7 @@ namespace net
 
         net::tx_op*     pop_tx_op();
         net::rx_page*   pop_rx_page();
-        uint32_t        handle_rx_page(net::rx_page* p);
+        uint64_t        handle_rx_page(net::rx_page* p);
 
         finterface(ipv4::addr ip_addr, uint16_t tx_mtu = 1500,
                    uint16_t rx_mtu = 1500);
